+++ eProject_HK2/database/migrations/2020_07_05_202949_create_game_producers_table.php
@@ -14,8 +14,8 @@ class CreateGameProducersTable extends Migration
     public function up()
     {
         Schema::create('game_producer', function (Blueprint $table) {
-            $table->integer('GAME_ID');
-            $table->integer('PRODUCER_ID');
+            $table->unsignedBigInteger('GAME_ID');
+            $table->unsignedBigInteger('PRODUCER_ID');
             $table->primary(array('GAME_ID', 'PRODUCER_ID'));
             $table->timestamps();
         });
